requirements.txt
argostranslate==1.9.6
coqui-tts==0.24.3
coqui-tts[ja]
Flask==3.1.0
Flask-Cors==5.0.0
numpy==1.26.4
requests==2.32.3
torch==2.5.1
transformers==4.46.2
urllib3==2.0.7
openai-whisper
torchaudio
scipy
librosa
soundfile
